Rename `classRestCom.py` -> `ig_rest_api.py`.
